hoàn thiện một ít giao diện material design đẹp

+++ DatPhatAcc/Models/DTO/TempTransDetailDTO.cs
@@ -13,8 +13,8 @@ namespace DatPhatAcc.Models.DTO
         public string GoodId { get; set; } = string.Empty;
         public string ShortName { get; set; } = string.Empty;
 
-        private int _quantity = 0;
-        public int Quantity
+        private decimal _quantity = 0;
+        public decimal Quantity
         {
             get => _quantity;
             set
